chore: rename dataType file
